requirements.txt
openai>=1.0.0
langchain>=0.1.0
langgraph>=0.0.40
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
redis>=5.0.0
python-dotenv>=1.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
moviepy>=1.0.0
elevenlabs>=0.2.0
google-api-python-client>=2.100.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0